Admin registration path

+++ src/main/java/com/louisngatale/registration_service/repositories/RolesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolesRepository extends JpaRepository<Roles,Integer> {
+public interface RolesRepository extends JpaRepository<Roles,Long> {
     Roles findByRole(String role);
 }
